+++ test/test_ImSim/test_MultiBand/test_multi_linear.py
@@ -2,10 +2,11 @@ __author__ = 'sibirrer'
 
 import numpy.testing as npt
 import pytest
+import numpy as np
 
 from lenstronomy.Data.imaging_data import ImageData
 from lenstronomy.Data.psf import PSF
-from lenstronomy.ImSim.MultiBand.multiband_multimodel import MultiBandMultiModel
+from lenstronomy.ImSim.MultiBand.multi_linear import MultiLinear
 import lenstronomy.Util.param_util as param_util
 from lenstronomy.LensModel.lens_model import LensModel
 from lenstronomy.LightModel.light_model import LightModel
@@ -34,7 +35,7 @@ class TestImageModel(object):
         data_class = ImageData(**kwargs_data)
         kwargs_psf = sim_util.psf_configure_simple(psf_type='GAUSSIAN', fwhm=fwhm, kernelsize=31, deltaPix=deltaPix,
                                                truncate=5)
-        psf_class = PSF(kwargs_psf)
+        psf_class = PSF(**kwargs_psf)
         # 'EXERNAL_SHEAR': external shear
         kwargs_shear = {'e1': 0.01, 'e2': 0.01}  # gamma_ext: shear strength, psi_ext: shear angel (in radian)
         phi, q = 0.2, 0.8
@@ -53,31 +54,34 @@ class TestImageModel(object):
         kwargs_sersic_ellipse = {'amp': 1., 'R_sersic': .6, 'n_sersic': 7, 'center_x': 0, 'center_y': 0,
                                  'e1': e1, 'e2': e2}
 
-        lens_light_model_list = ['SERSIC', 'SERSIC']
-        self.kwargs_lens_light = [kwargs_sersic, kwargs_sersic]
-        lens_light_model_class = LightModel(light_model_list=['SERSIC'])
-        source_model_list = ['SERSIC_ELLIPSE', 'SERSIC_ELLIPSE']
-        self.kwargs_source = [kwargs_sersic_ellipse, kwargs_sersic_ellipse]
-        source_model_class = LightModel(light_model_list=['SERSIC_ELLIPSE'])
+        lens_light_model_list = ['SERSIC']
+        self.kwargs_lens_light = [kwargs_sersic]
+        lens_light_model_class = LightModel(light_model_list=lens_light_model_list)
+        source_model_list = ['SERSIC_ELLIPSE']
+        self.kwargs_source = [kwargs_sersic_ellipse]
+        source_model_class = LightModel(light_model_list=source_model_list)
         self.kwargs_ps = [{'ra_source': 0.0001, 'dec_source': 0.0,
                            'source_amp': 1.}]  # quasar point source position in the source plane and intrinsic brightness
         point_source_class = PointSource(point_source_type_list=['SOURCE_POSITION'], fixed_magnification_list=[True])
-        kwargs_numerics = {'subgrid_res': 2, 'psf_subgrid': True}
+        kwargs_numerics = {'supersampling_factor': 2, 'supersampling_convolution': True, 'compute_mode': 'gaussian'}
         imageModel = ImageModel(data_class, psf_class, lens_model_class, source_model_class, lens_light_model_class,
                                 point_source_class, kwargs_numerics=kwargs_numerics)
-        image_sim = sim_util.simulate_simple(imageModel, self.kwargs_lens, [kwargs_sersic_ellipse],
-                                             [kwargs_sersic], self.kwargs_ps)
+        image_sim = sim_util.simulate_simple(imageModel, self.kwargs_lens, self.kwargs_source,
+                                         self.kwargs_lens_light, self.kwargs_ps)
         data_class.update_data(image_sim)
         kwargs_data['image_data'] = image_sim
         self.solver = LensEquationSolver(lensModel=lens_model_class)
-        kwargs_model_bool = {'index_source_light_model': [0], 'index_lens_light_model': [0]}
-        multi_band_list = [[kwargs_data, kwargs_psf, kwargs_numerics, kwargs_model_bool]]
-        self.imageModel = MultiBandMultiModel(multi_band_list, lens_model_class, source_model_list, lens_light_model_list, point_source_class)
+        multi_band_list = [[kwargs_data, kwargs_psf, kwargs_numerics]]
+        kwargs_model = {'lens_model_list': lens_model_list, 'source_light_model_list': source_model_list,
+                        'point_source_model_list': ['SOURCE_POSITION'], 'fixed_magnification_list': [True]}
+        self.imageModel = MultiLinear(multi_band_list, kwargs_model, likelihood_mask_list=None, compute_bool=None)
 
     def test_image_linear_solve(self):
         model, error_map, cov_param, param = self.imageModel.image_linear_solve(self.kwargs_lens, self.kwargs_source, self.kwargs_lens_light, self.kwargs_ps, inv_bool=False)
         chi2_reduced = self.imageModel._imageModel_list[0].reduced_chi2(model[0], error_map[0])
         npt.assert_almost_equal(chi2_reduced, 1, decimal=1)
+        chi2_reduced_list = self.imageModel.reduced_residuals(model_list=model, error_map_list=error_map)
+        npt.assert_almost_equal(np.sum(chi2_reduced_list[0]**2)/(100**2), 1, decimal=1)
 
     def test_likelihood_data_given_model(self):
         logL = self.imageModel.likelihood_data_given_model(self.kwargs_lens, self.kwargs_source, self.kwargs_lens_light, self.kwargs_ps, source_marg=False)
@@ -94,5 +98,3 @@ class TestImageModel(object):
 
 if __name__ == '__main__':
     pytest.main()
-
-