+++ test/test_ImSim/test_Numerics/test_convolution.py
@@ -52,24 +52,24 @@ class TestSubgridKernelConvolution(object):
 
     def test_convolve2d(self):
         #kernel_supersampled = kernel_util.subgrid_kernel(self.kernel, self.supersampling_factor, odd=True, num_iter=5)
-        subgrid_conv = SubgridKernelConvolution(self.kernel_sub, self.supersampling_factor, supersampling_size=None, convolution_type='fft')
+        subgrid_conv = SubgridKernelConvolution(self.kernel_sub, self.supersampling_factor, supersampling_kernel_size=None, convolution_type='fft')
         model_subgrid_conv = subgrid_conv.convolution2d(self.model_sub)
 
         supersampling_factor = 1
-        conv = SubgridKernelConvolution(self.kernel, supersampling_factor, supersampling_size=None,
+        conv = SubgridKernelConvolution(self.kernel, supersampling_factor, supersampling_kernel_size=None,
                                         convolution_type='fft')
         model_conv = conv.convolution2d(self.model)
         npt.assert_almost_equal(np.sum(model_subgrid_conv), np.sum(model_conv), decimal=1)
         npt.assert_almost_equal(model_subgrid_conv, model_conv, decimal=1)
 
         #kernel_supersampled = kernel_util.subgrid_kernel(self.kernel, self.supersampling_factor, odd=True, num_iter=5)
-        subgrid_conv_split = SubgridKernelConvolution(self.kernel_sub, self.supersampling_factor, supersampling_size=5,
+        subgrid_conv_split = SubgridKernelConvolution(self.kernel_sub, self.supersampling_factor, supersampling_kernel_size=5,
                                                 convolution_type='fft')
         model_subgrid_conv_split = subgrid_conv_split.convolution2d(self.model_sub)
         npt.assert_almost_equal(np.sum(model_subgrid_conv), np.sum(model_subgrid_conv_split), decimal=8)
         npt.assert_almost_equal(model_subgrid_conv, model_subgrid_conv_split, decimal=8)
 
-        subgrid_conv_split = SubgridKernelConvolution(self.kernel_sub, self.supersampling_factor, supersampling_size=2,
+        subgrid_conv_split = SubgridKernelConvolution(self.kernel_sub, self.supersampling_factor, supersampling_kernel_size=2,
                                                       convolution_type='fft')
         model_subgrid_conv_split = subgrid_conv_split.convolution2d(self.model_sub)
         npt.assert_almost_equal(np.sum(model_subgrid_conv), np.sum(model_subgrid_conv_split), decimal=5)